+++ backend/accounts/urls.py
@@ -1,15 +1,14 @@
 from django.urls import path
-from .views import PlaylistView, LikeListView, SignupAPIView, MyTokenObtainPairView, FollowAPIView, ProfileView, AccountsChangeView, PasswordChangeView, recommend_components, recommend_like, recommend_user, search_user
+from .views import get_like_list, get_play_list, signup, MyTokenObtainPairView, FollowAPIView, profile_view, change_accounts, recommend_components, recommend_like, recommend_user, search_user
 
 urlpatterns = [
     path('login/', MyTokenObtainPairView.as_view()),
-    path('playlist', PlaylistView.as_view()),
-    path('like', LikeListView.as_view()),
-    path('signup', SignupAPIView.as_view()),
     path('<int:user_pk>/follow', FollowAPIView.as_view()),
-    path('<int:user_pk>/profile', ProfileView.as_view()),
-    path('edit/<int:user_pk>/', AccountsChangeView.as_view()),
-    path('edit/<int:user_pk>/password', PasswordChangeView.as_view()),
+    path('playlist', get_play_list),
+    path('like', get_like_list),
+    path('signup', signup),
+    path('<int:user_pk>/profile', profile_view),
+    path('edit/<int:user_pk>/',change_accounts),
     path('recommend/components', recommend_components),
     path('recommend/user', recommend_user),
     path('recommend/like', recommend_like),
